Move Country under Resources::Public
