+++ spec/resources/fax_spec.rb
@@ -212,27 +212,6 @@ RSpec.describe Fax do
     end
   end
 
-  describe 'getting a list of supported countries' do
-    let(:action) { Fax.supported_countries }
-    let(:params) { {} }
-
-    around do |example|
-      VCR.use_cassette('resources/fax/supported_countries') do
-        example.run
-      end
-    end
-
-    it 'makes the request to Phaxio' do
-      expect_api_request :get, 'public/countries', params
-      action
-    end
-
-    it 'returns a collection of supported country resources' do
-      result = action
-      expect(result).to be_a(Phaxio::Resource::Collection)
-    end
-  end
-
   describe Fax::Reference do
     let(:fax_id) { 1234 }
